@@ -5,34 +5,34 @@
   "main": "index.js",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
-    "dev": "webpack-dev-server --config webpack.config.js --progress --env development",
+    "dev": "webpack-dev-server --config webpack.config.js -p --env development",
     "build": "webpack -p --env production"
   },
   "author": "",
   "license": "ISC",
   "dependencies": {
     "@aspnet/signalr": "1.0.3",
-    "babel-polyfill": "^6.26.0",
-    "chart.js": "2.7.0",
+    "chart.js": "2.7.3",
+    "core-js": "^2.5.7",
     "create-react-class": "15.6.2",
     "font-awesome": "4.7.0",
-    "line-awesome": "1.0.2",
-    "node-sass": "4.9.3",
     "griddle-react": "1.9.0",
     "highlight.js": "9.12.0",
     "izitoast": "1.3.0",
     "jquery": "3.2.1",
     "js-cookie": "2.2.0",
+    "line-awesome": "1.0.2",
     "materialize-css": "0.100.2",
     "moment": "2.19.3",
+    "node-sass": "4.9.3",
     "numeral": "2.0.6",
     "promise-polyfill": "6.0.2",
     "pubsub-js": "1.5.7",
-    "react": "16.5.2",
-    "react-chartjs-2": "2.6.2",
+    "react": "16.6.0",
+    "react-chartjs-2": "2.7.4",
     "react-color": "2.13.8",
     "react-debounce-input": "3.2.0",
-    "react-dom": "^16.4.2",
+    "react-dom": "^16.6.0",
     "react-interval": "^2.0.1",
     "react-materialize": "1.1.0",
     "react-redux": "5.0.7",
@@ -45,6 +45,7 @@
   "devDependencies": {
     "@babel/core": "7.1.2",
     "@babel/plugin-proposal-class-properties": "^7.1.0",
+    "@babel/plugin-syntax-dynamic-import": "^7.0.0",
     "@babel/polyfill": "^7.0.0",
     "@babel/preset-env": "7.1.0",
     "@babel/preset-react": "7.0.0",
@@ -57,6 +58,6 @@
     "uglifyjs-webpack-plugin": "2.0.1",
     "webpack": "4.21.0",
     "webpack-cli": "^3.1.2",
-    "webpack-dev-server": "^2.8.2"
+    "webpack-dev-server": "3.1.10"
   }
 }
