@@ -1,6 +1,6 @@
-import React from 'react';
-import {Link} from 'react-router-dom';
-import UdIcon from './ud-icon.jsx';
+import React,{Suspense} from 'react';
+
+const UdIconComponent = React.lazy(() => import('./ud-icon.jsx' /* webpackChunkName: "ud-icon" */))
 
 export default class UdLink extends React.Component {
     render() {
@@ -11,7 +11,11 @@ export default class UdLink extends React.Component {
         }
 
         if (this.props.icon) {
-            return <a href={this.props.url} target={target} className={this.props.className + " ud-link"} style={style}><UdIcon icon={this.props.icon}></UdIcon> {this.props.text}</a>
+            return <a href={this.props.url} target={target} className={this.props.className + " ud-link"} style={style}>
+            <Suspense fallback={<div>Loading...</div>}>
+                <UdIconComponent icon={this.props.icon}></UdIconComponent> {this.props.text}
+            </Suspense>
+            </a>
         }
         else {
             return <a href={this.props.url} target={target} className={this.props.className + " ud-link"} style={style}>{this.props.text}</a>
